Fix trust_proxy config parsing in Navigator
The trust_proxy field was defined in the Config struct but not being copied
from the YAML config during parsing. This caused trust_proxy to always be
false regardless of the YAML configuration value.

Changes:
- Added TrustProxy field to YAMLConfig.Server struct with yaml tag
- Added line to copy trust_proxy value in parseServerConfig()

This fixes the issue where Navigator would overwrite X-Forwarded-Host even
when trust_proxy: true was configured.